TextAnalyzer/requirements.txt
fastapi
uvicorn
python-multipart
pypdf2
python-docx